Change in the logic for scheduled status changes and small correction in interest calculation.

See merge request jala-university1/cohort-3/oficial-pt-desenvolvimento-de-software-2-iso-124.ga.t2.24.m2/se-o-b/Capstone/loans-module!27

+++ src/main/java/org/jala/university/presentation/controller/PaymentsController.java
@@ -142,7 +142,7 @@ public class PaymentsController {
         Integer numberOfInstallments = installmentsComboBox.getValue();
 
         if (amountBorrowed != null && numberOfInstallments != null && numberOfInstallments > 0) {
-            Double totalPayable = CalculationUtil.getTotalPayable(amountBorrowed, (double) numberOfInstallments);
+            Double totalPayable = CalculationUtil.getTotalPayable(amountBorrowed, numberOfInstallments);
             this.valueOfInstallments = totalPayable / numberOfInstallments;
 
             installmentValueLabel.setText(String.format("R$ %.2f per installment", valueOfInstallments));
